ListingDetailScreen now has special UI for delivery, showing 2 markers on map with appropriate zoom. Fixed Category layout in detailed. Removed unnecessary code.

+++ androidApp/src/main/java/com/example/plshelp/android/ui/screens/ListingsScreen.kt
@@ -302,7 +302,8 @@ fun ExpandableListingCard(
             val listingLat = listing.coord[0]
             val listingLon = listing.coord[1]
             val results = FloatArray(1)
-            Location.distanceBetween(currentLat, currentLon, listingLat.toDouble(), listingLon.toDouble(), results) // Ensure conversion to Double
+            Location.distanceBetween(currentLat, currentLon,
+                listingLat, listingLon, results)
             val distanceInMeters = results[0]
 
             when (displayMode) {
